requirements.txt
keyboard==0.13.5
pyautogui==0.9.54
Pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pynput==1.7.6
pystray==0.19.5
